source/processor/requirements.txt
ibm_cos_sdk==2.13.1
ibm_cos_sdk_core==2.13.1
pika==1.3.2
python-dotenv==1.0.0
python-multipart